Increase doinst.sh size to 32M
OK, this is probably overkill, but there are already some packages whose
doinst.sh files are bigger than 4M, so the initial comment doesn't apply
anymore. One such example is the qogir-icon-theme package, which in the
cut-down version without the ubuntu and manjaro variants creates more
than 64000 symlinks in its doinst.sh, as absurd as that seems.